Remove legacy automaton canvas file
