+++ lib/presentation/widgets/automaton_canvas/index.dart
@@ -1,0 +1,4 @@
+export 'automaton_canvas.dart';
+export 'automaton_painter.dart';
+export 'state_edit_dialog.dart';
+export 'transition_symbol_input.dart';
